refactor/#88 : del missionadd package
