+++ app/src/main/java/com/kiero/presentation/parent/screen/mission/directadd/component/MissionAwardInfo.kt
@@ -1,4 +1,4 @@
-package com.kiero.presentation.parent.screen.mission.directadd.component.missionadd
+package com.kiero.presentation.parent.screen.mission.directadd.component
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
